+++ CHANGELOG.md
@@ -61,6 +61,7 @@ Bug fixes:
 * Fixed enumeration issue in `ENV.{select, filter}`.
 * Fixed `Complex` and `Rational` should be frozen after initializing.
 * Fixed `printf` should raise error when not enough arguments for positional argument.
+* Removed "shadowing outer local variable" warning.
 
 Compatibility: